==> r/SouthWest/04_FSSGAM_fishspecies/03_predict-gam_SwC_fspecies_effectsplots.R <==
###
# Project: Gabby PhD Ch3 Abrolhos  Fish & habitat
# Data:    BRUV & BOSS fish MaxN
# Task:    Fish model prediction
# author:  Gabby Cummins
# date:    May 2024
##

rm(list=ls())

library(reshape2)
library(mgcv)
library(ggplot2)
library(viridis)
library(raster)
library(dplyr)
library(stringr)
library(terra)
library(sf)
library(gstat)
library(GlobalArchive)
library(patchwork)
library(png)
library(jpeg)
library(grid)

name <- "SouthWest"  # set study name

# read in
dat1 <- readRDS("data/staging/SwC/Southwest.fish.dat.maxn.rds")%>%
  #dplyr::mutate(reef =rock+inverts)%>%
  mutate(z = abs(z), name = scientific, scientific = paste(method,scientific,sep=".")) %>%
  #mutate(status = ifelse(is.na(status), "No-take", status)) %>%
  glimpse()

test1 <- dat1 %>%
  group_by(method)%>%
  summarise(z=mean(z), tpi = mean(tpi))

unique(dat1$scientific)

###RESIDUAL PLOTS FOR SOUTHWEST
dat1$method <- as.factor(dat1$method)

##Neeed to add fish pics here once found
p.bis <- readJPEG("data/images/swc/Pseudolabrus biserialis-3cm.jpg")
p.bis_grob <- rasterGrob(p.bis, width = unit(2.5, "cm"), height = unit(1.0, "cm"), interpolate = TRUE)

o.lin <- readJPEG("data/images/swc/Opthalmolepis lineolatus-3cm.jpg")
o.lin_grob <- rasterGrob(o.lin, width = unit(2.5, "cm"), height = unit(1.0, "cm"), interpolate = TRUE)

c.aur <- readJPEG("data/images/swc/Coris auricularis 3cm.jpg")
c.aur_grob <- rasterGrob(c.aur, width = unit(3.0, "cm"), height = unit(1.0, "cm"), interpolate = TRUE)

n.obl <- readJPEG("data/images/swc/Neatypus obliquus-3cmL.jpg")
n.obl_grob <- rasterGrob(n.obl, width = unit(2.75, "cm"), height = unit(1.5, "cm"), interpolate = TRUE)

##Pseudolabrus_biserialis BOSS BRUV
Pseudolabrus_biserialis_BOSSBRUV <- gam(number ~ s(z, k=3, bs = "cr", by = method) + s(tpi, k=3, bs = "cr", by = method) + method,
                                    data = dat1 %>% dplyr::filter(scientific %in% c("BOSS.Labridae Pseudolabrus biserialis","BRUV.Labridae Pseudolabrus biserialis")),
                                    family = tw())
summary(Pseudolabrus_biserialis_BOSSBRUV)


dat_total_pb <- dat1 %>%
  dplyr::filter(scientific %in% c("BOSS.Labridae Pseudolabrus biserialis","BRUV.Labridae Pseudolabrus biserialis"))

testdata_pb <- expand.grid(z = seq(min(dat_total_pb$z), max(dat_total_pb$z), length.out = 20),
                           tpi = mean(Pseudolabrus_biserialis_BOSSBRUV$model$tpi),
                           method = c("BOSS", "BRUV")) %>%
  distinct() %>%
  glimpse

P.biserlialis.fits <- predict.gam(Pseudolabrus_biserialis_BOSSBRUV, newdata = testdata_pb, type = 'response', se.fit = T)

predicts_pb_z <- testdata_pb %>%
  data.frame(P.biserlialis.fits) %>%
  group_by(z, method) %>%
  summarise(number = mean(fit), se.fit = mean(se.fit))  %>%
  ungroup() 

predicts_pb_z <- predicts_pb_z %>%
  mutate(method = factor(method, levels = c("BRUV", "BOSS")))

testdata1_pb <- expand.grid(method = c("BOSS", "BRUV"), 
                            z = mean(Pseudolabrus_biserialis_BOSSBRUV$model$z),
                            tpi = mean(Pseudolabrus_biserialis_BOSSBRUV$model$tpi)) %>%
  distinct() %>%
  glimpse() 

P.b.fits <- predict.gam(Pseudolabrus_biserialis_BOSSBRUV, newdata = testdata1_pb, type = 'response', se.fit = T)

predicts_pb_method <- testdata1_pb %>%
  data.frame(P.b.fits) %>%
  group_by(method) %>%
  summarise(number = mean(fit), se.fit = mean(se.fit)) %>%
  ungroup()

testdata2_pb <- expand.grid(tpi = seq(min(dat_total_pb$tpi), max(dat_total_pb$tpi), length.out = 20),
                            method = c("BOSS", "BRUV"),
                            z = mean(Pseudolabrus_biserialis_BOSSBRUV$model$z)) %>%
  distinct() %>%
  glimpse()

P.b.fits2 <- predict.gam(Pseudolabrus_biserialis_BOSSBRUV, newdata = testdata2_pb, type = 'response', se.fit = T)

predicts_pb_tpi <- testdata2_pb %>%
  data.frame(P.b.fits2) %>%
  group_by(tpi, method) %>%
  summarise(number = mean(fit), se.fit = mean(se.fit)) %>%
  ungroup()

predicts_pb_tpi <- predicts_pb_tpi %>%
  mutate(method = factor(method, levels = c("BRUV", "BOSS")))


#add custom text for fish name on plot
p.bis_text <- textGrob(label = expression(italic("P. biserialis")), x = 0, y = 0, 
                     just = "left", gp = gpar(col = "#000000", fontsize = 11))

# #add rug data ie raw data
pbis.dat <- dat1 %>%
  filter(name == "Labridae Pseudolabrus biserialis")

#Plot P biserialis residual abundance by depth (z)
gg_P_biserialis_z <- ggplot() + 
  # geom_point(data = dat1, aes(x = z, y = number, group=method, fill=method), alpha = 0.2, size = 1, show.legend = T) +
  geom_ribbon(data = predicts_pb_z, aes(x = z, ymin = number - se.fit, ymax = number + se.fit, fill = method, group = method), alpha = 0.2) +
  geom_line(data = predicts_pb_z, aes(x = z, y = number, group=method, colour=method))+
  geom_line(data = predicts_pb_z, aes(x = z, y = number - se.fit, group=method, colour=method), linetype = "dashed") +
  geom_line(data = predicts_pb_z, aes(x = z, y = number + se.fit, group=method, colour=method), linetype = "dashed") +
  geom_rug(data = pbis.dat, aes(x = z, colour = method), sides = "b", alpha = 0.5) +  # Rug plot on the bottom
  # geom_point(data = pbis.dat, x = )
  # geom_ribbon(data = predicts_total_z, aes(x = z, ymin = number - se.fit, ymax = number + se.fit, fill = method, group = method)) +
   theme_classic() +
  annotation_custom(p.bis_grob, xmin = -10, xmax = 0, ymin = -Inf, ymax = Inf) +
  annotation_custom(p.bis_text, xmin = -16, xmax = -6, ymin = 2, ymax = 2) + # Text annotation outside the plot
  theme_void() +
  coord_cartesian(clip = "off") +
  theme(plot.margin = margin(0.5, 0.5, 0.5, 2.0, "cm" ))+
  #ylim(0,50)+
  labs(x = "Depth", y = "Abundance", colour = "Method", fill = "Method") +
  scale_colour_manual(values = c("BRUV" = "#56B4E9", "BOSS" = "#E69F00")) +
  scale_fill_manual(values = c("BRUV" = "#56B4E9", "BOSS" = "#E69F00"))+
  theme(
    axis.line = element_line(colour = "black", linewidth = 0.5),
    axis.ticks = element_line(),
    axis.ticks.length = unit(3, "pt"),
    axis.title.x = element_text(size = 11),  # X axis title size
    axis.title.y = element_text(size = 11, angle = 90, vjust = 1.5, margin = margin(4,4,0,5)),  # Y axis title size
    axis.text.x = element_text(size = 10, margin = margin(0, 0, 3, 0)),   # X axis text size
    axis.text.y = element_text(size = 10, margin = margin(0, 3, 0, 0)),   # Y axis text size
    legend.title = element_text(size = 11),  # Legend title size
    legend.text = element_text(size = 9) # Legend text size
  )
gg_P_biserialis_z

##Plot P nebulosa residual abundance by reef
gg_P_biserialis_tpi <- ggplot() + 
  geom_ribbon(data = predicts_pb_tpi, aes(x = tpi, ymin = number - se.fit, ymax = number + se.fit, fill = method, group = method), alpha = 0.2) +
  # geom_point(data = dat1, aes(x = z, y = number, group=method, fill=method), alpha = 0.2, size = 1, show.legend = T) +
  geom_line(data = predicts_pb_tpi, aes(x = tpi, y = number, group=method, colour=method))+
  geom_line(data = predicts_pb_tpi, aes(x = tpi, y = number - se.fit, group=method, colour=method), linetype = "dashed") +
  geom_line(data = predicts_pb_tpi, aes(x = tpi, y = number + se.fit, group=method, colour=method), linetype = "dashed") +
  geom_rug(data = pbis.dat, aes(x = tpi, colour = method), sides = "b", alpha = 0.5) +  # Rug plot on the bottom
  # geom_point(data = pbis.dat, aes(x = reef, y = number))+
  theme_classic() +
  #ylim(0,50)+
  labs(x = "TPI", y = "Abundance", colour = "Method", fill = "Method") +
  scale_colour_manual(values = c("BRUV" = "#56B4E9", "BOSS" = "#E69F00")) +
  scale_fill_manual(values = c("BRUV" = "#56B4E9", "BOSS" = "#E69F00"))+
  theme(
    axis.title.x = element_text(size = 11),  # X axis title size
    axis.title.y = element_text(size = 11),  # Y axis title size
    axis.text.x = element_text(size = 10),   # X axis text size
    axis.text.y = element_text(size = 10),   # Y axis text size
    legend.title = element_text(size = 11),  # Legend title size
    legend.text = element_text(size = 9)    # Legend text size
  )
gg_P_biserialis_tpi

# test <- dat1 %>% dplyr::filter(scientific %in% c("BOSS.Labridae Ophthalmolepis lineolatus","BRUV.Labridae Ophthalmolepis lineolatus"))
# hist(test$number)

## Same as above but for Opthalmolepsis lineolatus
Opthalmolepsis_lineolatus_BOSSBRUV <- gam(number ~ s(z, k=3, bs = "cr", by = method) + s(tpi, k=3, bs = "cr", by = method) + method,
                                    data = dat1 %>% dplyr::filter(scientific %in% c("BOSS.Labridae Ophthalmolepis lineolatus","BRUV.Labridae Ophthalmolepis lineolatus")),
                                    family = tw())
summary(Opthalmolepsis_lineolatus_BOSSBRUV)

plot(Opthalmolepsis_lineolatus_BOSSBRUV)

dat_total_ol <- dat1 %>%
  dplyr::filter(scientific %in% c("BOSS.Labridae Ophthalmolepis lineolatus", "BRUV.Labridae Ophthalmolepis lineolatus"))

testdata_ol <- expand.grid(z = seq(min(dat_total_ol$z), max(dat_total_ol$z), length.out = 20),
                           tpi = mean(Opthalmolepsis_lineolatus_BOSSBRUV$model$tpi),
                           method = c("BOSS", "BRUV")) %>%
  distinct() %>%
  glimpse

O.lineolatus.fits <- predict.gam(Opthalmolepsis_lineolatus_BOSSBRUV, newdata = testdata_ol, type = 'response', se.fit = T)

predicts_ol_z <- testdata_ol %>%
  data.frame(O.lineolatus.fits) %>%
  group_by(z, method) %>%
  summarise(number = mean(fit), se.fit = mean(se.fit))  %>%
  ungroup() 

#swap so BRUV is on top of the legend
predicts_ol_z <- predicts_ol_z %>%
  mutate(method = factor(method, levels = c("BRUV", "BOSS")))

# testdata1_ol <- expand.grid(method = c("BOSS", "BRUV"), 
#                             z = mean(Opthalmolepsis_lineolatus_BOSSBRUV$model$z),
#                             reef = mean(Opthalmolepsis_lineolatus_BOSSBRUV$model$reef)) %>%
#   distinct() %>%
#   glimpse() 
# 
# O.l.fits <- predict.gam(Opthalmolepsis_lineolatus_BOSSBRUV, newdata = testdata1_ol, type = 'response', se.fit = T)
# 
# predicts_ol_method <- testdata1_ol %>%
#   data.frame(O.l.fits) %>%
#   group_by(method) %>%
#   summarise(number = mean(fit), se.fit = mean(se.fit)) %>%
#   ungroup()

testdata2_ol <- expand.grid(tpi = seq(min(dat_total_ol$tpi), max(dat_total_ol$tpi), length.out = 20),
                            method = c("BOSS", "BRUV"),
                            z = mean(Opthalmolepsis_lineolatus_BOSSBRUV$model$z)) %>%
  distinct() %>%
  glimpse()

O.l.fits2 <- predict.gam(Opthalmolepsis_lineolatus_BOSSBRUV, newdata = testdata2_ol, type = 'response', se.fit = T)

predicts_ol_tpi <- testdata2_ol %>%
  data.frame(O.l.fits2) %>%
  group_by(tpi, method) %>%
  summarise(number = mean(fit), se.fit = mean(se.fit)) %>%
  ungroup()

#swap so BRUV is on top of the legend
predicts_ol_tpi <- predicts_ol_tpi %>%
  mutate(method = factor(method, levels = c("BRUV", "BOSS")))

#add custom text for fish name
 o.lin_text <- textGrob(label = expression(italic("O. lineolatus")), x = 0, y = 0, 
                        just = "left", gp = gpar(col = "#000000", fontsize = 11))

#add rug data ie raw data
olin.dat <- dat1 %>%
  filter(name == "Labridae Ophthalmolepis lineolatus")

#Plot Ophthalmolepis lineolatus residual abundance by depth (z)
gg_O_lineolatus_z <- ggplot() + 
  # geom_point(data = dat1, aes(x = z, y = number, group=method, fill=method), alpha = 0.2, size = 1, show.legend = T) +
  geom_ribbon(data = predicts_ol_z, aes(x = z, ymin = number - se.fit, ymax = number + se.fit, fill = method, group = method), alpha = 0.2) +
  geom_line(data = predicts_ol_z, aes(x = z, y = number, group=method, colour=method))+
  geom_line(data = predicts_ol_z, aes(x = z, y = number - se.fit, group=method, colour=method), linetype = "dashed") +
  geom_line(data = predicts_ol_z, aes(x = z, y = number + se.fit, group=method, colour=method), linetype = "dashed") +
  geom_rug(data = olin.dat, aes(x = z, colour = method), sides = "b", alpha = 0.5) + 
  # geom_ribbon(data = predicts_total_z, aes(x = z, ymin = number - se.fit, ymax = number + se.fit, fill = method, group = method)) +
  theme_classic() +
  annotation_custom(o.lin_grob, xmin = -10, xmax = 0, ymin = -Inf, ymax = Inf) +
  annotation_custom(o.lin_text, xmin = -17, xmax = -7, ymin = 2.5, ymax = 2.5) +
  theme_void() +
  coord_cartesian(clip = "off") +
  theme(plot.margin = margin(0.5, 0.5, 0.5, 3.5, "cm" ))+
  #ylim(0,50)+
  labs(x = "Depth", y = "Abundance", colour = "Method", fill = "Method") +
  scale_colour_manual(values = c("BRUV" = "#56B4E9", "BOSS" = "#E69F00")) +
  scale_fill_manual(values = c("BRUV" = "#56B4E9", "BOSS" = "#E69F00"))+
  theme(
    axis.line = element_line(colour = "black", linewidth = 0.5),
    axis.ticks = element_line(),
    axis.ticks.length = unit(3, "pt"),
    axis.title.x = element_text(size = 11),  # X axis title size
    axis.title.y = element_text(size = 11, angle = 90, vjust = 1.5, margin = margin(4,4,0,5)),  # Y axis title size
    axis.text.x = element_text(size = 10, margin = margin(0, 0, 3, 0)),   # X axis text size
    axis.text.y = element_text(size = 10, margin = margin(0, 3, 0, 0)),   # Y axis text size
    legend.title = element_text(size = 11),  # Legend title size
    legend.text = element_text(size = 9) # Legend text size
  )
gg_O_lineolatus_z

#Plot Ophthalmolepis lineolatus residual abundance by reef
gg_O_lineolatus_tpi <- ggplot() + 
  geom_ribbon(data = predicts_ol_tpi, aes(x = tpi, ymin = number - se.fit, ymax = number + se.fit, fill = method, group = method), alpha = 0.2) +
  # geom_point(data = dat1, aes(x = z, y = number, group=method, fill=method), alpha = 0.2, size = 1, show.legend = T) +
  geom_line(data = predicts_ol_tpi, aes(x = tpi, y = number, group=method, colour=method))+
  geom_line(data = predicts_ol_tpi, aes(x = tpi, y = number - se.fit, group=method, colour=method), linetype = "dashed") +
  geom_line(data = predicts_ol_tpi, aes(x = tpi, y = number + se.fit, group=method, colour=method), linetype = "dashed") +
  geom_rug(data = olin.dat, aes(x = tpi, colour = method), sides = "b", alpha = 0.5) + 
  # geom_point(data = olin.dat, aes(x = reef, y = number))+
  theme_classic() +
    ylim(0,1)+
  labs(x = "TPI", y = "Abundance", colour = "Method", fill = "Method") +
  scale_colour_manual(values = c("BRUV" = "#56B4E9", "BOSS" = "#E69F00")) +
  scale_fill_manual(values = c("BRUV" = "#56B4E9", "BOSS" = "#E69F00"))+
  theme(
    axis.line = element_line(colour = "black", linewidth = 0.5),
    axis.ticks = element_line(),
    axis.ticks.length = unit(3, "pt"),
    axis.title.x = element_text(size = 11),  # X axis title size
    axis.title.y = element_text(size = 11, angle = 90, vjust = 1.5, margin = margin(4,4,0,5)),  # Y axis title size
    axis.text.x = element_text(size = 10, margin = margin(0, 0, 3, 0)),   # X axis text size
    axis.text.y = element_text(size = 10, margin = margin(0, 3, 0, 0)),   # Y axis text size
    legend.title = element_text(size = 11),  # Legend title size
    legend.text = element_text(size = 9) # Legend text size
  )
gg_O_lineolatus_tpi

## Same as above but for Coris auricularis
Coris_auricularis_BOSSBRUV <- gam(number ~ s(z, k=3, bs = "cr", by = method) + s(tpi, k=3, bs = "cr", by = method) + method,
                                  data = dat1 %>% dplyr::filter(scientific %in% c("BOSS.Labridae Coris auricularis","BRUV.Labridae Coris auricularis")),
                                  family = tw())
summary(Coris_auricularis_BOSSBRUV)

dat_total_ca <- dat1 %>%
  dplyr::filter(scientific %in% c("BOSS.Labridae Coris auricularis", "BRUV.Labridae Coris auricularis"))

testdata_ca <- expand.grid(z = seq(min(dat_total_ca$z), max(dat_total_ca$z), length.out = 20),
                           tpi = mean(Coris_auricularis_BOSSBRUV$model$tpi),
                           method = c("BOSS", "BRUV")) %>%
  distinct() %>%
  glimpse

C.auricularis.fits <- predict.gam(Coris_auricularis_BOSSBRUV, newdata = testdata_ca, type = 'response', se.fit = T)

predicts_ca_z <- testdata_ca %>%
  data.frame(C.auricularis.fits) %>%
  group_by(z, method) %>%
  summarise(number = mean(fit), se.fit = mean(se.fit))  %>%
  ungroup() 

#swap so BRUV is on top of the legend
predicts_ca_z <- predicts_ca_z %>%
  mutate(method = factor(method, levels = c("BRUV", "BOSS")))

testdata1_ca <- expand.grid(method = c("BOSS", "BRUV"), 
                            z = mean(Coris_auricularis_BOSSBRUV$model$z),
                            tpi = mean(Coris_auricularis_BOSSBRUV$model$tpi)) %>%
  distinct() %>%
  glimpse() 

C.a.fits <- predict.gam(Coris_auricularis_BOSSBRUV, newdata = testdata1_ca, type = 'response', se.fit = T)

predicts_ca_method <- testdata1_ca %>%
  data.frame(C.a.fits) %>%
  group_by(method) %>%
  summarise(number = mean(fit), se.fit = mean(se.fit)) %>%
  ungroup()

testdata2_ca <- expand.grid(tpi = seq(min(dat_total_ca$tpi), max(dat_total_ca$tpi), length.out = 20),
                            method = c("BOSS", "BRUV"),
                            z = mean(Coris_auricularis_BOSSBRUV$model$z)) %>%
  distinct() %>%
  glimpse()

C.a.fits2 <- predict.gam(Coris_auricularis_BOSSBRUV, newdata = testdata2_ca, type = 'response', se.fit = T)

predicts_ca_tpi <- testdata2_ca %>%
  data.frame(C.a.fits2) %>%
  group_by(tpi, method) %>%
  summarise(number = mean(fit), se.fit = mean(se.fit)) %>%
  ungroup()

#swap so BRUV is on top of the legend
predicts_ca_tpi <- predicts_ca_tpi %>%
  mutate(method = factor(method, levels = c("BRUV", "BOSS")))

# #add custom text for fish name
c.aur_text <- textGrob(label = expression(italic("C. auricularis")), x = 0, y = 0, 
                        just = "left", gp = gpar(col = "#000000", fontsize = 11))

#add rug data ie raw data
caur.dat <- dat1 %>%
  filter(name == "Labridae Coris auricularis")

#Plot C auricularis residual abundance by depth (z)
gg_C_auricularis_z <- ggplot() + 
  # geom_point(data = dat1, aes(x = z, y = number, group=method, fill=method), alpha = 0.2, size = 1, show.legend = T) +
  geom_ribbon(data = predicts_ca_z, aes(x = z, ymin = number - se.fit, ymax = number + se.fit, fill = method, group = method), alpha=0.2) +
  geom_line(data = predicts_ca_z, aes(x = z, y = number, group=method, colour=method))+
  geom_line(data = predicts_ca_z, aes(x = z, y = number - se.fit, group=method, colour=method), linetype = "dashed") +
  geom_line(data = predicts_ca_z, aes(x = z, y = number + se.fit, group=method, colour=method), linetype = "dashed") +
  geom_rug(data = caur.dat, aes(x = z, colour = method), sides = "b", alpha = 0.5) +
  # geom_ribbon(data = predicts_total_z, aes(x = z, ymin = number - se.fit, ymax = number + se.fit, fill = method, group = method)) +
  theme_classic() +
  annotation_custom(c.aur_grob, xmin = -17, xmax = 7, ymin = -Inf, ymax = Inf) +
  annotation_custom(c.aur_text, xmin = -20, xmax = -10, ymin = 7, ymax = 7) +
  theme_void() +
  coord_cartesian(clip = "off") +
  theme(plot.margin = margin(0.5, 0.5, 0.5, 3.5, "cm" ))+
  #ylim(0,50)+
  labs(x = "Depth", y = "Abundance", colour = "Method", fill = "Method") +
  scale_colour_manual(values = c("BRUV" = "#56B4E9", "BOSS" = "#E69F00")) +
  scale_fill_manual(values = c("BRUV" = "#56B4E9", "BOSS" = "#E69F00"))+
  theme(
    axis.line = element_line(colour = "black", linewidth = 0.5),
    axis.ticks = element_line(),
    axis.ticks.length = unit(3, "pt"),
    axis.title.x = element_text(size = 11),  # X axis title size
    axis.title.y = element_text(size = 11, angle = 90, vjust = 1.5, margin = margin(4,4,0,5)),  # Y axis title size
    axis.text.x = element_text(size = 10, margin = margin(0, 0, 3, 0)),   # X axis text size
    axis.text.y = element_text(size = 10, margin = margin(0, 3, 0, 0)),   # Y axis text size
    legend.title = element_text(size = 11),  # Legend title size
    legend.text = element_text(size = 9) # Legend text size
  )
gg_C_auricularis_z

# # Save gg_total_z as a PNG file
# ggsave(filename = "plots/Abrolhos/BOSSBRUV_Cauricularis_depth_residualplot.png", 
#        plot = gg_C_auricularis_z, 
#        width = 6, 
#        height = 4, 
#        dpi = 600, 
#        units = "in")

#Plot C.auricularis residual abundance by reef
gg_C_auricularis_tpi <- ggplot() + 
  geom_ribbon(data = predicts_ca_tpi, aes(x = tpi, ymin = number - se.fit, ymax = number + se.fit, fill = method, group = method), alpha = 0.2) +
  # geom_point(data = dat1, aes(x = z, y = number, group=method, fill=method), alpha = 0.2, size = 1, show.legend = T) +
  geom_line(data = predicts_ca_tpi, aes(x = tpi, y = number, group=method, colour=method))+
  geom_line(data = predicts_ca_tpi, aes(x = tpi, y = number - se.fit, group=method, colour=method), linetype = "dashed") +
  geom_line(data = predicts_ca_tpi, aes(x = tpi, y = number + se.fit, group=method, colour=method), linetype = "dashed") +
  geom_rug(data = caur.dat, aes(x = tpi, colour = method), sides = "b", alpha = 0.5) +
  theme_classic() +
  #ylim(0,50)+
  labs(x = "TPI", y = "Abundance", colour = "Method", fill = "Method") +
  scale_colour_manual(values = c("BRUV" = "#56B4E9", "BOSS" = "#E69F00")) +
  scale_fill_manual(values = c("BRUV" = "#56B4E9", "BOSS" = "#E69F00"))+
  theme(
    axis.line = element_line(colour = "black", linewidth = 0.5),
    axis.ticks = element_line(),
    axis.ticks.length = unit(3, "pt"),
    axis.title.x = element_text(size = 11),  # X axis title size
    axis.title.y = element_text(size = 11, angle = 90, vjust = 1.5, margin = margin(4,4,0,5)),  # Y axis title size
    axis.text.x = element_text(size = 10, margin = margin(0, 0, 3, 0)),   # X axis text size
    axis.text.y = element_text(size = 10, margin = margin(0, 3, 0, 0)),   # Y axis text size
    legend.title = element_text(size = 11),  # Legend title size
    legend.text = element_text(size = 9) # Legend text size
  )
gg_C_auricularis_tpi

# # Save gg_total_reef as a PNG file
# ggsave(filename = "plots/Abrolhos/BOSSBRUV_Cauricularis_reef_residualplot.png", 
#        plot = gg_C_auricularis_reef, 
#        width = 6, 
#        height = 4, 
#        dpi = 600, 
#        units = "in")

## Same as above but for Neatypus obliquus 
Neatypus_obliquus_BOSSBRUV <- gam(number ~ s(z, k=3, bs = "cr", by = method) + s(tpi, k=3, bs = "cr", by = method) + method,
                                          data = dat1 %>% dplyr::filter(scientific %in% c("BOSS.Scorpididae Neatypus obliquus","BRUV.Scorpididae Neatypus obliquus")),
                                          family = tw())
summary(Neatypus_obliquus_BOSSBRUV)

plot(Neatypus_obliquus_BOSSBRUV)

dat_total_no <- dat1 %>%
  dplyr::filter(scientific %in% c("BOSS.Scorpididae Neatypus obliquus","BRUV.Scorpididae Neatypus obliquus"))

testdata_no <- expand.grid(z = seq(min(dat_total_no$z), max(dat_total_no$z), length.out = 20),
                           tpi = mean(Opthalmolepsis_lineolatus_BOSSBRUV$model$tpi),
                           method = c("BOSS", "BRUV")) %>%
  distinct() %>%
  glimpse

N.obliquus.fits <- predict.gam(Neatypus_obliquus_BOSSBRUV, newdata = testdata_ol, type = 'response', se.fit = T)

predicts_no_z <- testdata_no %>%
  data.frame(N.obliquus.fits) %>%
  group_by(z, method) %>%
  summarise(number = mean(fit), se.fit = mean(se.fit))  %>%
  ungroup() 

#swap so BRUV is on top of the legend
predicts_no_z <- predicts_no_z %>%
  mutate(method = factor(method, levels = c("BRUV", "BOSS")))

# testdata1_no <- expand.grid(method = c("BOSS", "BRUV"), 
#                             z = mean(Neatypus_obliquus_BOSSBRUV$model$z),
#                             reef = mean(Neatypus_obliquus_BOSSBRUV$model$reef)) %>%
#   distinct() %>%
#   glimpse() 
# 
# N.o.fits <- predict.gam(Neatypus_obliquus_BOSSBRUV, newdata = testdata1_no, type = 'response', se.fit = T)
# 
# predicts_no_method <- testdata1_no %>%
#   data.frame(N.o.fits) %>%
#   group_by(method) %>%
#   summarise(number = mean(fit), se.fit = mean(se.fit)) %>%
#   ungroup()

testdata2_no <- expand.grid(tpi = seq(min(dat_total_no$tpi), max(dat_total_no$tpi), length.out = 20),
                            method = c("BOSS", "BRUV"),
                            z = mean(Neatypus_obliquus_BOSSBRUV$model$z)) %>%
  distinct() %>%
  glimpse()

N.o.fits2 <- predict.gam(Neatypus_obliquus_BOSSBRUV, newdata = testdata2_no, type = 'response', se.fit = T)

predicts_no_tpi <- testdata2_no %>%
  data.frame(N.o.fits2) %>%
  group_by(tpi, method) %>%
  summarise(number = mean(fit), se.fit = mean(se.fit)) %>%
  ungroup()

#swap so BRUV is on top of the legend
predicts_no_tpi <- predicts_no_tpi %>%
  mutate(method = factor(method, levels = c("BRUV", "BOSS")))

#add custom text for fish name
 n.obl_text <- textGrob(label = expression(italic("N. obliquus")), x = 0, y = 0, 
                       just = "left", gp = gpar(col = "#000000", fontsize = 11))

#add rug data ie raw data
nobl.dat <- dat1 %>%
  filter(name == "Scorpididae Neatypus obliquus") 

#Plot Neatypus obliquus residual abundance by depth (z)
gg_N_obliquus_z <- ggplot() + 
  # geom_point(data = dat1, aes(x = z, y = number, group=method, fill=method), alpha = 0.2, size = 1, show.legend = T) +
  geom_ribbon(data = predicts_no_z, aes(x = z, ymin = number - se.fit, ymax = number + se.fit, fill = method, group = method), alpha = 0.2) +
  geom_line(data = predicts_no_z, aes(x = z, y = number, group=method, colour=method))+
  geom_line(data = predicts_no_z, aes(x = z, y = number - se.fit, group=method, colour=method), linetype = "dashed") +
  geom_line(data = predicts_no_z, aes(x = z, y = number + se.fit, group=method, colour=method), linetype = "dashed") +
  geom_rug(data = nobl.dat, aes(x = z, colour = method), sides = "b", alpha = 0.5) + 
  # geom_ribbon(data = predicts_total_z, aes(x = z, ymin = number - se.fit, ymax = number + se.fit, fill = method, group = method)) +
  theme_classic() +
  annotation_custom(n.obl_grob, xmin = -13, xmax = -3, ymin = -Inf, ymax = Inf) +
  annotation_custom(n.obl_text, xmin = -20, xmax = -10, ymin = 3.5, ymax = 3.5) +
  theme_void() +
  coord_cartesian(clip = "off") +
  theme(plot.margin = margin(0.5, 0.5, 0.5, 3.5, "cm" ))+
  #ylim(0,50)+
  labs(x = "Depth", y = "Abundance", colour = "Method", fill = "Method") +
  scale_colour_manual(values = c("BRUV" = "#56B4E9", "BOSS" = "#E69F00")) +
  scale_fill_manual(values = c("BRUV" = "#56B4E9", "BOSS" = "#E69F00"))+
  theme(
    axis.line = element_line(colour = "black", linewidth = 0.5),
    axis.ticks = element_line(),
    axis.ticks.length = unit(3, "pt"),
    axis.title.x = element_text(size = 11),  # X axis title size
    axis.title.y = element_text(size = 11, angle = 90, vjust = 1.5, margin = margin(4,4,0,5)),  # Y axis title size
    axis.text.x = element_text(size = 10, margin = margin(0, 0, 3, 0)),   # X axis text size
    axis.text.y = element_text(size = 10, margin = margin(0, 3, 0, 0)),   # Y axis text size
    legend.title = element_text(size = 11),  # Legend title size
    legend.text = element_text(size = 9) # Legend text size
  )
gg_N_obliquus_z


#Plot Ophthalmolepis lineolatus residual abundance by TPI
gg_N_obliquus_tpi <- ggplot() + 
  geom_ribbon(data = predicts_no_tpi, aes(x = tpi, ymin = number - se.fit, ymax = number + se.fit, fill = method, group = method), alpha = 0.2) +
  # geom_point(data = dat1, aes(x = z, y = number, group=method, fill=method), alpha = 0.2, size = 1, show.legend = T) +
  geom_line(data = predicts_no_tpi, aes(x = tpi, y = number, group=method, colour=method))+
  geom_line(data = predicts_no_tpi, aes(x = tpi, y = number - se.fit, group=method, colour=method), linetype = "dashed") +
  geom_line(data = predicts_no_tpi, aes(x = tpi, y = number + se.fit, group=method, colour=method), linetype = "dashed") +
  geom_rug(data = nobl.dat, aes(x = tpi, colour = method), sides = "b", alpha = 0.5) + 
  # geom_point(data = olin.dat, aes(x = reef, y = number))+
  theme_classic() +
  # ylim(0,1)+
  labs(x = "TPI", y = "Abundance", colour = "Method", fill = "Method") +
  scale_colour_manual(values = c("BRUV" = "#56B4E9", "BOSS" = "#E69F00")) +
  scale_fill_manual(values = c("BRUV" = "#56B4E9", "BOSS" = "#E69F00"))+
  theme(
    axis.line = element_line(colour = "black", linewidth = 0.5),
    axis.ticks = element_line(),
    axis.ticks.length = unit(3, "pt"),
    axis.title.x = element_text(size = 11),  # X axis title size
    axis.title.y = element_text(size = 11, angle = 90, vjust = 1.5, margin = margin(4,4,0,5)),  # Y axis title size
    axis.text.x = element_text(size = 10, margin = margin(0, 0, 3, 0)),   # X axis text size
    axis.text.y = element_text(size = 10, margin = margin(0, 3, 0, 0)),   # Y axis text size
    legend.title = element_text(size = 11),  # Legend title size
    legend.text = element_text(size = 9) # Legend text size
  )
gg_N_obliquus_tpi

#join altogether
p_swc <- gg_P_biserialis_z + gg_P_biserialis_tpi + gg_O_lineolatus_z + gg_O_lineolatus_tpi + gg_C_auricularis_z  + gg_C_auricularis_tpi + gg_N_obliquus_z + gg_N_obliquus_tpi + plot_layout(ncol = 2)
p_swc

#save
ggsave(filename = "plots/SwC/BOSSBRUV_Southwest_effectsplots_TPI.png", 
       plot = p_swc, 
       width = 12, 
       height = 16, 
       dpi = 600, 
       units = "in")

#depth only
depth_swc <- gg_P_biserialis_z + gg_O_lineolatus_z + gg_C_auricularis_z + gg_N_obliquus_z + plot_layout(ncol = 1)
depth_swc

#save
ggsave(filename = "plots/SwC/BOSSBRUV_Southwest_effectsplots_depthonly.png", 
       plot = depth_swc, 
       width = 7, 
       height = 16, 
       dpi = 600, 
       units = "in")

#Note below file taken from output of Ch2
preddf <- readRDS("data/spatial/rasters/raw bathymetry/SouthWest_spatial_habitat_predictions.rds") %>%
  # dplyr::rename(reef = pinverts.fit + pmacroal)%>% ###DONT NEED
  ga.clean.names()%>%
  mutate(z =abs(z))

#creating boss and bruv dfs
preddf_boss <- preddf %>%
  dplyr::mutate(method = "BOSS")

preddf_bruv <- preddf %>%
  dplyr::mutate(method = "BRUV")

preddf_new <- bind_rows(preddf_boss, preddf_bruv)

# predict, rasterise and plot ### CHANGE TO se.fit = TRUE fo individual fish species!
predicted_fish <- cbind(preddf_new, 
                        "p_P_biserialis" = predict(Pseudolabrus_biserialis_BOSSBRUV, preddf_new, type = "response", se.fit = T),
                        "p_O_lineolatus" = predict(Opthalmolepsis_lineolatus_BOSSBRUV, preddf_new, type = "response", se.fit = T),
                        "p_C_auricularis" = predict(Coris_auricularis_BOSSBRUV, preddf_new, type = "response", se.fit = T),
                        "p_N_obliquus" = predict(Neatypus_obliquus_BOSSBRUV, preddf_new, type = "response", se.fit = T))

prasts_boss <- predicted_fish %>%
  dplyr::filter(method %in% "BOSS") %>%
  dplyr::select(x, y, z, starts_with("p_")) %>%
  rast(crs = "epsg:4326")
plot(prasts_boss)

boss_fish <- as.data.frame(prasts_boss, xy = T)

prasts_bruv <- predicted_fish %>%
  dplyr::filter(method %in% "BRUV") %>%
  dplyr::select(x, y, z, starts_with("p_")) %>%
  rast(crs = "epsg:4326")
plot(prasts_bruv)

bruv_fish <- as.data.frame(prasts_bruv, xy = T)

saveRDS(boss_fish, paste0("outputs/SwC/fish/", name, "_BOSS_predicted-fish.RDS"))

saveRDS(bruv_fish, paste0("outputs/SwC/fish/", name, "_BRUV_predicted-fish.RDS"))
